requirements.txt
discord.py
PyGithub
requests
